+++ pages/_document.tsx
@@ -1,0 +1,14 @@
+import { NextPage } from 'next'
+import { Head, Html, Main, NextScript } from 'next/document'
+
+const Document: NextPage = () => (
+  <Html lang='de-DE' prefix='og: http://ogp.me/ns#'>
+    <Head />
+    <body>
+      <Main />
+      <NextScript />
+    </body>
+  </Html>
+)
+
+export default Document