+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added [admin user creation script](prisma/createAdminUser.mjs)
 - Added NextAuth to project
 - Added session handling
+- Added [custom next Document](pages/_document.tsx)
 
 ### Changed